+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Current changes
 
+- Fixed a bug in calculating the node order when connecting a multiple output node to different nodes per output.
+
 ## v0.5.0
 
 Large breaking changes in this version